# docs/api/openapi-v1.yaml
openapi: 3.1.0
info:
  title: Finanzas App API
  version: 1.0.0
  description: API v1 para app financiera Chile-first (CLP).
servers:
  - url: https://api.finanzasapp.cl/v1
security:
  - bearerAuth: []
...
paths:
...
  /goals:
    get:
      tags: [Goals]
      summary: Listar metas
      responses:
        '200':
          description: Metas
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/SavingGoal'
    post:
      tags: [Goals]
      summary: Crear meta
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateGoalRequest'
      responses:
        '201':
          description: Meta creada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SavingGoal'
  /goals/{goalId}/contributions:
    post:
      tags: [Goals]
      summary: Registrar aporte a meta
      parameters:
        - in: path
          name: goalId
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amountClp]
              properties:
                amountClp:
                  type: integer
                  minimum: 1
                note:
                  type: string
      responses:
        '201':
          description: Aporte registrado
  /budgets:
    get:
      tags: [Budgets]
      summary: Listar presupuestos mensuales
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Presupuestos
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/Budget'
    put:
      tags: [Budgets]
      summary: Upsert de presupuesto por categoria y mes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpsertBudgetRequest'
      responses:
        '200':
          description: Presupuesto guardado
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Budget'
  /reports/dashboard:
    get:
      tags: [Reports]
      summary: KPIs de inicio
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Resumen mensual
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DashboardReport'
  /reports/monthly-comparison:
    get:
      tags: [Reports]
      summary: Comparativo de meses
      parameters:
        - in: query
          name: fromMonth
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
        - in: query
          name: toMonth
          required: true
          schema:
            type: string
            pattern: '^\d{4}-\d{2}$'
      responses:
        '200':
          description: Serie comparativa
  /reports/export.csv:
    get:
      tags: [Reports]
      summary: Exporta movimientos a CSV
      parameters:
        - in: query
          name: month
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Archivo CSV
          content:
            text/csv:
              schema:
                type: string
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    CreateGoalRequest:
      type: object
      required: [name, targetAmountClp, monthlyContributionClp]
      properties:
        name: { type: string }
        targetAmountClp: { type: integer, minimum: 1 }
        monthlyContributionClp: { type: integer, minimum: 1 }
        targetDate: { type: string, format: date }
    SavingGoal:
      type: object
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        targetAmountClp: { type: integer }
        monthlyContributionClp: { type: integer }
        currentAmountClp: { type: integer }
        status: { type: string, enum: [active, paused, completed, cancelled] }
    UpsertBudgetRequest:
      type: object
      required: [categoryId, year, month, amountLimitClp]
      properties:
        categoryId: { type: string, format: uuid }
        year: { type: integer, minimum: 2020 }
        month: { type: integer, minimum: 1, maximum: 12 }
        amountLimitClp: { type: integer, minimum: 1 }
    Budget:
      type: object
      properties:
        id: { type: string, format: uuid }
        categoryId: { type: string, format: uuid }
        year: { type: integer }
        month: { type: integer }
        amountLimitClp: { type: integer }
    DashboardReport:
      type: object
      properties:
        month: { type: string }
        incomeTotalClp: { type: integer }
        expenseTotalClp: { type: integer }
        savingTotalClp: { type: integer }
        overspendPercent: { type: number }